feat: move middleware to generis, deprecate code and update doc
